delete prediction req + make prediction req(w/o calling ml)

+++ backend/db_init.sql
@@ -16,4 +16,4 @@ create table predictions (
     create_time bigint,
     predict_time bigint,
     predicted_price double precision,
-    user_id integer references users (id));
+    user_id bigint references users (id));
